requirements.txt
langgraph
langchain-openai
langchain-groq
langchain-community
langchain_google_genai
faiss-cpu

ipykernel
python-dotenv
pypdf
cryptography
pillow
langgraph-checkpoint-sqlite

aiosqlite

streamlit
fastapi



-e .